+++ app/modules/auction/application/handlers/created_auction_handler.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 
 from app.modules.auction.domain.events.auction_events import AuctionCreatedEvent
 from app.modules.auction.infrastructure.persistence.auction_read_model import AuctionReadModel
@@ -16,8 +17,8 @@ class AuctionCreatedHandler:
             user_id=event.payload["user_id"],
             title=event.payload["title"],
             description=event.payload["description"],
-            start_price=float(event.payload["start_price"]),
-            minimum_increment=float(event.payload["minimum_increment"]),
+            start_price=Decimal(event.payload["start_price"]),
+            minimum_increment=Decimal(event.payload["minimum_increment"]),
             images=event.payload["images"],
         )
         await self.read_repository.save(auction)
